feat: add InvalidIdException to CompanyController and EmployeeController

+++ src/main/java/com/rest/springbootemployee/controller/CompanyController.java
@@ -4,8 +4,10 @@ import com.rest.springbootemployee.controller.dto.CompanyRequest;
 import com.rest.springbootemployee.controller.dto.CompanyResponse;
 import com.rest.springbootemployee.controller.mapper.CompanyMapper;
 import com.rest.springbootemployee.entity.Company;
+import com.rest.springbootemployee.exception.InvalidIdException;
 import com.rest.springbootemployee.service.CompanyService;
 import com.rest.springbootemployee.entity.Employee;
+import org.bson.types.ObjectId;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -30,11 +32,17 @@ public class CompanyController {
 
     @GetMapping("/{id}")
     public CompanyResponse getById(@PathVariable String id) {
+        if(!ObjectId.isValid(id)){
+            throw new InvalidIdException();
+        }
         return companyMapper.toResponse(companyService.findById(id));
     }
 
     @GetMapping("/{id}/employees")
     public List<Employee> getEmployees(@PathVariable String id) {
+        if(!ObjectId.isValid(id)){
+            throw new InvalidIdException();
+        }
         return companyService.getEmployees(id);
     }
 
@@ -54,6 +62,9 @@ public class CompanyController {
 
     @PutMapping("/{id}")
     public CompanyResponse update(@PathVariable String id, @RequestBody CompanyRequest companyRequest) {
+        if(!ObjectId.isValid(id)){
+            throw new InvalidIdException();
+        }
         Company company = companyMapper.toEntity(companyRequest);
         Company updatedCompany = companyService.update(id, company);
         return companyMapper.toResponse(updatedCompany);
@@ -62,6 +73,9 @@ public class CompanyController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteCompany(@PathVariable String id) {
+        if(!ObjectId.isValid(id)){
+            throw new InvalidIdException();
+        }
         companyService.delete(id);
     }
 }
